pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.4.57",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "python-binance>=1.0.27",
    "python-socketio[client]>=5.12.1",
    "requests>=2.32.3",
    "websocket-client>=1.8.0",
    "websockets>=14.2",
]
